pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["emptyCan"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
"discord.py" = "^2.1.1"
"dislash.py" = "^1.4.9"
discord-buttons-plugin = "^0.1.2"
requests = "^2.28.2"
gspread = "^5.7.2"
oauth2client = "^4.1.3"
xds-protos = "^0.0.11"
Pillow = "^9.4.0"
emoji = "^2.2.0"
PyNaCl = "^1.5.0"
pyglet = "^2.0.5"
discord = "^2.2.2"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"